Agile and XP practices : Overview of Extreme Programming
The previous chapter gave us an outline of what agile software development is about. However, it didn’t tell us
exactly what to do. It gave us some platitudes and goals, but it gave us little in the way of real direction. This chapter
corrects that.
The Practices of Extreme Programming
Extreme programming is the most famous of the agile methods. It is made up of a set of simple, yet interdependent
practices. These practices work together to form a whole that is greater than its parts. We shall briefly consider that
whole in this chapter, and examine some of the parts in chapters to come.